requirements-eval.txt
scipy
torchvision
nltk
inflection
pycocoevalcap
pycocotools
tqdm
scikit-learn
black
mypy
pylint
pytest
requests